backend/requirements.txt
fastapi
uvicorn
python-dotenv
anthropic
python-multipart
httpx
